refactor to add license package
