+++ src/main/java/edu/harvard/iq/dataverse/TermsOfUseAndAccess.java
@@ -15,6 +15,8 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToOne;
 
+import edu.harvard.iq.dataverse.license.License;
+
 /**
  *
  *
