+++ src/main/java/edu/harvard/iq/dataverse/license/License.java
@@ -1,4 +1,4 @@
-package edu.harvard.iq.dataverse;
+package edu.harvard.iq.dataverse.license;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -10,6 +10,9 @@ import javax.persistence.NamedQuery;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.persistence.UniqueConstraint;
+
+import edu.harvard.iq.dataverse.TermsOfUseAndAccess;
+
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
